Move the operation status message trait to the smartrest crate
