+++ plugins/c8y_configuration_plugin/src/main.rs
@@ -1,7 +1,6 @@
 mod config;
 mod download;
 mod error;
-mod smartrest;
 mod upload;
 
 use crate::config::PluginConfig;
